start merge mbo_scan and lazy-array
