+++ mbo_utilities/file_io.py
@@ -13,12 +13,8 @@ import dask.array as da
 from tifffile import TiffFile
 
 from . import log
-from ._parsing import _make_json_serializable
-from .metadata import is_raw_scanimage, get_metadata
-from .phasecorr import nd_windowed, ALL_PHASECORR_METHODS
-from .scanreader import scans, utils
-from .scanreader.multiroi import ROI
-from .util import subsample_array
+from .lazy_array import Scan_MBO
+from .metadata import is_raw_scanimage
 
 try:
     from zarr import open as zarr_open
@@ -371,7 +367,7 @@ def read_scan(
 
     Returns
     -------
-    Scan_MBO
+    mbo_utilities.lazy_array.Scan_MBO
         A scan object with metadata and lazily loaded data. Raises FileNotFoundError
         if no files match the specified path(s).
 
@@ -410,614 +406,6 @@ def read_scan(
     scan.read_data(filenames, dtype=dtype)
     return scan
 
-class ZarrScanView2:
-    def __init__(self, zarr_array, ys, xs, oys, oxs, roi=None, metadata={}):
-        self.z = zarr_array  # zarr or dask backed
-        self.yslices = ys
-        self.xslices = xs
-        self.oyslices = oys
-        self.oxslices = oxs
-        self.num_rois = len(ys)
-        self.shape = self.z.shape
-        self.roi = roi
-
-    def __getitem__(self, key):
-        if not isinstance(key, tuple):
-            key = (key,)
-        t_key, z_key, y_key, x_key = (key + (slice(None),) * 4)[:4]
-
-        if self.roi is not None:
-            if isinstance(self.roi, (list, tuple)):
-                return tuple(
-                    self.z[t_key, z_key, self.yslices[r - 1], self.xslices[r - 1]]
-                    for r in self.roi
-                )
-            elif self.roi == 0:
-                return tuple(
-                    self.z[t_key, z_key, self.yslices[r], self.xslices[r]]
-                    for r in range(self.num_rois)
-                )
-            elif isinstance(self.roi, int):
-                return self.z[t_key, z_key, self.yslices[self.roi - 1], self.xslices[self.roi - 1]]
-            else:
-                raise ValueError(f"Invalid ROI type: {type(self.roi)}. Expected int, list, or None.")
-        else:
-            print("Concatenating ROI's")
-            t_len = 1 if isinstance(t_key, int) else len(np.arange(*t_key.indices(self.shape[0])))
-            z_len = 1 if isinstance(z_key, int) else len(np.arange(*z_key.indices(self.shape[1])))
-            assembled = np.zeros(
-                (t_len, z_len,
-                 self.oyslices[0].stop - self.oyslices[0].start,
-                 max(s.stop for s in self.oxslices)),
-                dtype=self.z.dtype
-            )
-            for r in range(self.num_rois):
-                xs, ys = self.xslices[r], self.yslices[r]
-                oxs, oys = self.oxslices[r], self.oyslices[r]
-                print(f"ROI {r}: ys={ys}, xs={xs}")
-                print(f"ROI {r}: oys={oys}, oxs={oxs}")
-                print(f"ROI {r}: input zarr shape={self.z.shape}")
-
-                sub = self.z[t_key, z_key, ys, xs]
-                if sub.ndim == 3:
-                    sub = sub[:, np.newaxis, :, :]
-                elif sub.ndim == 2:
-                    sub = sub[np.newaxis, np.newaxis, :, :]
-                h, w = sub.shape[-2:]
-                assembled[..., oys.start:oys.start + h, oxs.start:oxs.start + w] = sub
-            return assembled
-
-class ZarrScanView:
-    def __init__(self, zarr_array, ys, xs, oys, oxs, roi=None, metadata=None):
-        self.z = zarr_array
-        self.yslices = ys
-        self.xslices = xs
-        self.oyslices = oys
-        self.oxslices = oxs
-        self.num_rois = len(ys)
-        self.shape = zarr_array.shape
-        self.roi = roi
-        self.metadata = metadata or {}
-
-    def __getitem__(self, key):
-        key = (key,) if not isinstance(key, tuple) else key
-        t_key, z_key, *_ = (key + (slice(None),) * 4)[:4]
-
-        if self.roi is not None:
-            def extract_roi(r):
-                return self.z[t_key, z_key, self.yslices[r], self.xslices[r]]
-            if self.roi == 0:
-                return tuple(extract_roi(r) for r in range(self.num_rois))
-            if isinstance(self.roi, (list, tuple)):
-                return tuple(extract_roi(r - 1) for r in self.roi)
-            if isinstance(self.roi, int):
-                return extract_roi(self.roi - 1)
-            raise ValueError(f"Invalid ROI type: {type(self.roi)}")
-
-        t_len = 1 if isinstance(t_key, int) else len(np.arange(*t_key.indices(self.shape[0])))
-        z_len = 1 if isinstance(z_key, int) else len(np.arange(*z_key.indices(self.shape[1])))
-        h = self.oyslices[0].stop - self.oyslices[0].start
-        w = max(s.stop for s in self.oxslices)
-        print((h, w))
-        assembled = np.zeros((t_len, z_len, h, w), dtype=self.z.dtype)
-
-        for ys, xs, oys, oxs in zip(self.yslices, self.xslices, self.oyslices, self.oxslices):
-            sub = self.z[t_key, z_key, ys, xs]
-            if sub.ndim == 3:
-                sub = sub[:, np.newaxis, :, :]
-            elif sub.ndim == 2:
-                sub = sub[np.newaxis, np.newaxis, :, :]
-            h, w = sub.shape[-2:]
-            print("--")
-            print((h, w))
-            assembled[..., oys.start:oys.start + h, oxs.start:oxs.start + w] = sub
-        return assembled
-
-
-class Scan_MBO(scans.ScanMultiROI):
-    """
-    A subclass of ScanMultiROI that ignores the num_fields dimension
-    and reorders the output to [time, z, x, y].
-    """
-
-    def __init__(
-            self,
-            roi: int | Sequence[int] | None = None,
-            fix_phase: bool = True,
-            phasecorr_method: str = "frame",
-            border: int | tuple[int, int, int, int] = 3,
-            upsample: int =5,
-            max_offset: int = 4,
-    ):
-        super().__init__(join_contiguous=True)
-        self._metadata = {} # set when pages are read
-        self.roi = roi  # alias
-        self._roi = roi
-        self._fix_phase = fix_phase
-        self._phasecorr_method = phasecorr_method
-        self.border: int | tuple[int, int, int, int] = border
-        self.max_offset: int = max_offset
-        self.upsample: int = upsample
-        self.pbar = None
-        self.show_pbar = False
-        self._offset = 0.0
-        self.use_zarr = True
-
-        # Debugging toggles
-        self.debug_flags = {
-            "frame_idx": True,
-            "roi_array_shape": False,
-            "phase_offset": False,
-        }
-        self.logger = logger
-        self.logger.info(
-            f"Initializing MBO Scan with parameters:\n"
-            f"roi: {roi}, "
-            f"fix_phase: {fix_phase}, "
-            f"phasecorr_method: {phasecorr_method}, "
-            f"border: {border}, "
-            f"upsample: {upsample}, "
-            f"max_offset: {max_offset}"
-        )
-
-    def save_fsspec(self, filenames):
-        base_dir = Path(filenames[0]).parent
-
-        combined_json_path = base_dir / "combined_refs.json"
-
-        if combined_json_path.is_file():
-            # delete it, its cheap to create
-            logger.debug(f"Removing existing combined reference file: {combined_json_path}")
-            combined_json_path.unlink()
-
-        print(f"Generating combined kerchunk reference for {len(filenames)} files…")
-        combined_refs = _multi_tiff_to_fsspec(tif_files=filenames, base_dir=base_dir)
-
-        with open(combined_json_path, "w") as _f:
-            json.dump(combined_refs, _f)
-
-        print(f"Combined kerchunk reference written to {combined_json_path}")
-        self.reference = combined_json_path
-        return combined_json_path
-
-    def as_dask(self):
-        """
-        Convert the current scan data to a Dask array.
-        This will create a Dask array in the same directory as the reference file.
-        """
-        if not HAS_ZARR:
-            raise ImportError("Zarr is not installed. Please install it to use this method.")
-        if not Path(self.reference).is_file():
-            raise FileNotFoundError(
-                f"Reference file {self.reference} does not exist. "
-                "Please call save_fsspec() first."
-            )
-        return da.from_zarr(
-            FsspecStore(ReferenceFileSystem(str(self.reference))),
-            chunks=CHUNKS,
-        )
-
-    def as_zarr(self):
-        """
-        Convert the current scan data to a Zarr array.
-        This will create a Zarr store in the same directory as the reference file.
-        """
-        if not HAS_ZARR:
-            raise ImportError("Zarr is not installed. Please install it to use this method.")
-        if not Path(self.reference).is_file():
-            raise FileNotFoundError(
-                f"Reference file {self.reference} does not exist. "
-                "Please call save_fsspec() first."
-            )
-        return zarr_open(
-            FsspecStore(ReferenceFileSystem(str(self.reference))),
-            mode="r",
-        )
-
-    def read_data(self, filenames, dtype=np.int16):
-        filenames = expand_paths(filenames)
-        self.save_fsspec(filenames)
-        super().read_data(filenames, dtype)
-        self._metadata = get_metadata(self.tiff_files[0].filehandle.path)  # from the file
-        self._metadata.update({"si": _make_json_serializable(self.tiff_files[0].scanimage_metadata)})
-        self._rois = self._create_rois()
-        self.fields = self._create_fields()
-        if self.join_contiguous:
-            self._join_contiguous_fields()
-
-    @property
-    def metadata(self):
-        md = self._metadata.copy()
-        md.update({
-            "fix_phase": self.fix_phase,
-            "phasecorr_method": self.phasecorr_method,
-            "offset": self.offset,
-            "border": self.border,
-            "upsample": self.upsample,
-            "max_offset": self.max_offset,
-        })
-        return md
-
-    @metadata.setter
-    def metadata(self, value):
-        self._metadata.update(value)
-
-    @property
-    def rois(self):
-        """ROI's hold information about the size, position and shape of the ROIs."""
-        return self._rois
-
-    @property
-    def offset(self):
-        return self._offset
-
-    @offset.setter
-    def offset(self, value: float | np.ndarray):
-        """
-        Set the phase offset for phase correction.
-        If value is a scalar, it applies the same offset to all frames.
-        If value is an array, it must match the number of frames.
-        """
-        if isinstance(value, int):
-            self._offset = float(value)
-        self._offset = value
-
-    @property
-    def phasecorr_method(self):
-        """
-        Get the current phase correction method.
-        Options are 'subpix' or 'mean'.
-        """
-        return self._phasecorr_method
-
-    @phasecorr_method.setter
-    def phasecorr_method(self, value: str):
-        """
-        Set the phase correction method.
-        Options are 'two_step', 'subpix', or 'crosscorr'.
-        """
-        if value not in ALL_PHASECORR_METHODS:
-            raise ValueError(
-                f"Unsupported phase correction method: {value}. "
-                f"Supported methods are: {ALL_PHASECORR_METHODS}"
-            )
-        self._phasecorr_method = value
-
-    @property
-    def fix_phase(self):
-        """
-        Get whether phase correction is applied.
-        If True, phase correction is applied to the data.
-        """
-        return self._fix_phase
-
-    @fix_phase.setter
-    def fix_phase(self, value: bool):
-        """
-        Set whether to apply phase correction.
-        If True, phase correction is applied to the data.
-        """
-        if not isinstance(value, bool):
-            raise ValueError("do_phasecorr must be a boolean value.")
-        self._fix_phase = value
-
-    @property
-    def roi(self):
-        """
-        Get the current ROI index.
-        If roi is None, returns -1 to indicate no specific ROI.
-        """
-        return self._roi
-
-    @roi.setter
-    def roi(self, value):
-        """
-        Set the current ROI index.
-        If value is None, sets roi to -1 to indicate no specific ROI.
-        """
-        self._roi = value
-
-    @property
-    def num_rois(self) -> int:
-        return len(self.rois)
-
-    @property
-    def xslices(self):
-        return self.fields[0].xslices
-
-    @property
-    def yslices(self):
-        return self.fields[0].yslices
-
-    @property
-    def output_xslices(self):
-        return self.fields[0].output_xslices
-
-    @property
-    def output_yslices(self):
-        return self.fields[0].output_yslices
-
-    def _read_pages(self, frames, chans, yslice=slice(None), xslice=slice(None), **kwargs):
-        C = self.num_channels
-        pages = [f * C + c for f in frames for c in chans]
-
-        H = len(utils.listify_index(yslice, self._page_height))
-        W = len(utils.listify_index(xslice, self._page_width))
-
-        if getattr(self, "use_zarr", False):
-            zarray = self.as_zarr()
-            buf = np.empty((len(pages), H, W), dtype=self.dtype)
-            for i, page in enumerate(pages):
-                f, c = divmod(page, C)
-                buf[i] = zarray[f, c, yslice, xslice]
-
-            if self.fix_phase:
-                self.logger.debug(f"Applying phase correction with strategy: {self.phasecorr_method}")
-                buf, self.offset = nd_windowed(
-                    buf,
-                    method=self.phasecorr_method,
-                    upsample=self.upsample,
-                    max_offset=self.max_offset,
-                    border=self.border,
-                )
-            return buf.reshape(len(frames), len(chans), H, W)
-
-        # TIFF path
-        buf = np.empty((len(pages), H, W), dtype=self.dtype)
-        start = 0
-        for tf in self.tiff_files:
-            end = start + len(tf.pages)
-            idxs = [i for i, p in enumerate(pages) if start <= p < end]
-            if not idxs:
-                start = end
-                continue
-
-            frame_idx = [pages[i] - start for i in idxs]
-            chunk = tf.asarray(key=frame_idx)[..., yslice, xslice]
-
-            if self.fix_phase:
-                self.logger.debug(f"Applying phase correction with strategy: {self.phasecorr_method}")
-                corrected, self.offset = nd_windowed(
-                    chunk,
-                    method=self.phasecorr_method,
-                    upsample=self.upsample,
-                    max_offset=self.max_offset,
-                    border=self.border,
-                )
-                buf[idxs] = corrected
-            else:
-                buf[idxs] = chunk
-            start = end
-
-        return buf.reshape(len(frames), len(chans), H, W)
-
-    def _read_pages2(
-        self, frames, chans, yslice=slice(None), xslice=slice(None), **kwargs
-    ):
-        C = self.num_channels
-        pages = [f * C + c for f in frames for c in chans]
-
-        H = len(utils.listify_index(yslice, self._page_height))
-        W = len(utils.listify_index(xslice, self._page_width))
-        if getattr(self, "use_zarr", False):
-            zarray = self.as_zarr()
-            buf = np.empty((len(frames), len(chans), H, W), dtype=self.dtype)
-            for fi, f in enumerate(frames):
-                for ci, c in enumerate(chans):
-                    buf[fi, ci] = zarray[f, c, yslice, xslice]
-            return buf
-
-        buf = np.empty((len(pages), H, W), dtype=self.dtype)
-        start = 0
-        for tf in self.tiff_files:
-            end = start + len(tf.pages)
-            idxs = [i for i, p in enumerate(pages) if start <= p < end]
-            if not idxs:
-                start = end
-                continue
-
-            frame_idx = [pages[i] - start for i in idxs]
-            chunk = tf.asarray(key=frame_idx)[..., yslice, xslice]
-
-            if self.fix_phase:
-                self.logger.debug(f"Applying phase correction with strategy: {self.phasecorr_method}")
-                corrected, self.offset = nd_windowed(
-                    chunk,
-                    method=self.phasecorr_method,
-                    upsample=self.upsample,
-                    max_offset=self.max_offset,
-                    border=self.border,
-                )
-                buf[idxs] = corrected
-            else:
-                buf[idxs] = chunk
-            start = end
-        return buf.reshape(len(frames), len(chans), H, W)
-
-    def _read_from_zarr(self, t_key, z_key):
-        if self.roi is not None:
-            def extract(r):
-                return self.zarr_array[t_key, z_key, self.yslices[r], self.xslices[r]]
-
-            if self.roi == 0:
-                return tuple(extract(r) for r in range(self.num_rois))
-            if isinstance(self.roi, (list, tuple)):
-                return tuple(extract(r - 1) for r in self.roi)
-            if isinstance(self.roi, int):
-                return extract(self.roi - 1)
-            raise ValueError(f"Invalid ROI type: {type(self.roi)}")
-
-        t_len = 1 if isinstance(t_key, int) else len(np.arange(*t_key.indices(self.shape[0])))
-        z_len = 1 if isinstance(z_key, int) else len(np.arange(*z_key.indices(self.shape[1])))
-        h = self.output_yslices[0].stop - self.output_yslices[0].start
-        w = max(s.stop for s in self.output_xslices)
-        assembled = np.zeros((t_len, z_len, h, w), dtype=self.zarr_array.dtype)
-
-        for ys, xs, oys, oxs in zip(self.yslices, self.xslices, self.output_yslices, self.output_xslices):
-            sub = self.zarr_array[t_key, z_key, ys, xs]
-            if sub.ndim == 3:
-                sub = sub[:, np.newaxis, :, :]
-            elif sub.ndim == 2:
-                sub = sub[np.newaxis, np.newaxis, :, :]
-            h, w = sub.shape[-2:]
-            assembled[..., oys.start:oys.start + h, oxs.start:oxs.start + w] = sub
-        return assembled
-
-    def __getitem__(self, key):
-        if not isinstance(key, tuple):
-            key = (key,)
-        t_key, z_key, _, _ = tuple(_convert_range_to_slice(k) for k in key) + (slice(None),) * (4 - len(key))
-        frames = utils.listify_index(t_key, self.num_frames)
-        chans = utils.listify_index(z_key, self.num_channels)
-        if not frames or not chans:
-            return np.empty(0)
-
-        logger.debug(
-            f"Phase-corrected: {self.fix_phase}/{self.phasecorr_method},"
-            f" channels: {chans},"
-            f" roi: {self.roi}",
-        )
-        out = self.process_rois(frames, chans)
-
-        squeeze = []
-        if isinstance(t_key, int):
-            squeeze.append(0)
-        if isinstance(z_key, int):
-            squeeze.append(1)
-        if squeeze:
-            if isinstance(out, tuple):
-                out = tuple(np.squeeze(x, axis=tuple(squeeze)) for x in out)
-            else:
-                out = np.squeeze(out, axis=tuple(squeeze))
-        return out
-
-    def process_rois(self, frames, chans):
-        if self.roi is not None:
-            if isinstance(self.roi, list):
-                return tuple(self.process_single_roi(roi_idx - 1, frames, chans) for roi_idx in self.roi)
-            elif self.roi == 0:
-                return tuple(self.process_single_roi(roi_idx, frames, chans) for roi_idx in range(self.num_rois))
-            elif isinstance(self.roi, int):
-                return self.process_single_roi(self.roi - 1, frames, chans)
-        else:
-            H_out, W_out = self.field_heights[0], self.field_widths[0]
-            out = np.zeros((len(frames), len(chans), H_out, W_out), dtype=self.dtype)
-            for roi_idx in range(self.num_rois):
-                roi_data = self.process_single_roi(roi_idx, frames, chans)
-                oys, oxs = self.fields[0].output_yslices[roi_idx], self.fields[0].output_xslices[roi_idx]
-                out[:, :, oys, oxs] = roi_data
-            return out
-
-    def process_single_roi(self, roi_idx, frames, chans):
-        return self._read_pages(
-            frames,
-            chans,
-            yslice=self.fields[0].yslices[roi_idx],
-            xslice=self.fields[0].xslices[roi_idx],
-        )
-
-    @property
-    def total_frames(self):
-        return sum(len(tf.pages) // self.num_channels for tf in self.tiff_files)
-
-    @property
-    def num_planes(self):
-        """LBM alias for num_channels."""
-        return self.num_channels
-
-    def min(self):
-        """
-        Returns the minimum value of the first tiff page.
-        """
-        page = self.tiff_files[0].pages[0]
-        return np.min(page.asarray())
-
-    def max(self):
-        """
-        Returns the maximum value of the first tiff page.
-        """
-        page = self.tiff_files[0].pages[0]
-        return np.max(page.asarray())
-
-    @property
-    def shape(self):
-        """Shape is relative to the current ROI."""
-        if self.roi is not None:
-            if not isinstance(self.roi, (list, tuple)):
-                if self.roi > 0:
-                    s = self.fields[0].output_xslices[self.roi - 1]
-                    width = s.stop - s.start
-                    return (
-                        self.total_frames,
-                        self.num_channels,
-                        self.field_heights[0],
-                        width,
-                    )
-        # roi = None, or a list/tuple indicates the shape should be relative to the full FOV
-        return (
-            self.total_frames,
-            self.num_channels,
-            self.field_heights[0],
-            self.field_widths[0],
-        )
-
-    @property
-    def shape_full(self):
-        return (
-            self.total_frames,
-            self.num_channels,
-            self.field_heights[0],
-            self.field_widths[0],
-        )
-
-    @property
-    def ndim(self):
-        return 4
-
-    @property
-    def size(self):
-        return (
-            self.num_frames
-            * self.num_channels
-            * self.field_heights[0]
-            * self.field_widths[0]
-        )
-
-    @property
-    def scanning_depths(self):
-        """
-        We override this because LBM should always be at a single scanning depth.
-        """
-        return [0]
-
-    def _create_rois(self):
-        """
-        Create scan rois from the configuration file. Override the base method to force
-        ROI's that have multiple 'zs' to a single depth.
-        """
-        try:
-            roi_infos = self.tiff_files[0].scanimage_metadata["RoiGroups"]["imagingRoiGroup"]["rois"]
-        except KeyError:
-            raise RuntimeError("This file is not a raw-scanimage tiff or is missing tiff.scanimage_metadata.")
-        roi_infos = roi_infos if isinstance(roi_infos, list) else [roi_infos]
-
-        # discard empty/malformed ROIs
-        roi_infos = list(filter(lambda r: isinstance(r["zs"], (int, float, list)), roi_infos))
-
-        # LBM uses a single depth that is not stored in metadata,
-        # so force this to be 0.
-        for roi_info in roi_infos:
-            roi_info["zs"] = [0]
-
-        rois = [ROI(roi_info) for roi_info in roi_infos]
-        return rois
-
-    def __array__(self):
-        """
-        Convert the scan data to a NumPy array.
-        Calculate the size of the scan and subsample to keep under memory limits.
-        """
-        return subsample_array(self, ignore_dims=[-1, -2, -3])
 
 def get_files(
     base_dir, str_contains="", max_depth=1, sort_ascending=True, exclude_dirs=None
